# src/app/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Loyalty Demo API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/webhooks/roller/booking": {
      "post": {
        "summary": "Consume a ROLLER booking webhook and acknowledge non-retryable payload issues.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "sendDate": {
                    "type": "string",
                    "minLength": 1
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "Booking"
                    ]
                  },
                  "eventType": {
                    "type": "string",
                    "enum": [
                      "Created",
                      "Updated",
                      "Cancelled",
                      "Deleted"
                    ]
                  },
                  "eventDate": {
                    "type": "string",
                    "minLength": 1
                  },
                  "data": {
                    "type": "object",
                    "properties": {
                      "bookingReference": {
                        "anyOf": [
                          {
                            "type": "string",
                            "minLength": 1
                          },
                          {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          }
                        ]
                      },
                      "uniqueId": {
                        "type": "string",
                        "minLength": 1
                      },
                      "customerId": {
                        "anyOf": [
                          {
                            "type": "string",
                            "minLength": 1
                          },
                          {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          }
                        ]
                      },
                      "email": {
                        "type": "string"
                      },
                      "firstName": {
                        "type": "string",
                        "minLength": 1
                      },
                      "lastName": {
                        "type": "string",
                        "minLength": 1
                      },
                      "name": {
                        "type": "string",
                        "minLength": 1
                      },
                      "phone": {
                        "type": "string",
                        "minLength": 1
                      },
                      "venue": {
                        "type": "string",
                        "minLength": 1
                      },
                      "startsAt": {
                        "type": "string",
                        "minLength": 1
                      },
                      "ticketCount": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      },
                      "status": {
                        "type": "string",
                        "minLength": 1
                      },
                      "items": {
                        "anyOf": [
                          {
                            "type": "object",
                            "properties": {
                              "quantity": {
                                "type": "integer",
                                "minimum": -9007199254740991,
                                "maximum": 9007199254740991
                              },
                              "bookingDate": {
                                "type": "string",
                                "minLength": 1
                              },
                              "sessionStartTime": {
                                "type": "string",
                                "minLength": 1
                              },
                              "sessionEndTime": {
                                "type": "string",
                                "minLength": 1
                              }
                            },
                            "additionalProperties": {}
                          },
                          {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "quantity": {
                                  "type": "integer",
                                  "minimum": -9007199254740991,
                                  "maximum": 9007199254740991
                                },
                                "bookingDate": {
                                  "type": "string",
                                  "minLength": 1
                                },
                                "sessionStartTime": {
                                  "type": "string",
                                  "minLength": 1
                                },
                                "sessionEndTime": {
                                  "type": "string",
                                  "minLength": 1
                                }
                              },
                              "additionalProperties": {}
                            }
                          }
                        ]
                      }
                    },
                    "additionalProperties": {}
                  }
                },
                "required": [
                  "id",
                  "sendDate",
                  "type",
                  "eventType",
                  "eventDate",
                  "data"
                ],
                "additionalProperties": {}
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Webhook accepted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "ok"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/webhooks/roller/signed-waiver": {
      "post": {
        "summary": "Consume a ROLLER signed waiver webhook and acknowledge non-retryable payload issues.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "sendDate": {
                    "type": "string",
                    "minLength": 1
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "SignedWaiver"
                    ]
                  },
                  "eventType": {
                    "type": "string",
                    "enum": [
                      "Created",
                      "Updated",
                      "Cancelled",
                      "Deleted"
                    ]
                  },
                  "eventDate": {
                    "type": "string",
                    "minLength": 1
                  },
                  "data": {
                    "minItems": 1,
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "signedWaiverId": {
                          "anyOf": [
                            {
                              "type": "string",
                              "minLength": 1
                            },
                            {
                              "type": "integer",
                              "minimum": -9007199254740991,
                              "maximum": 9007199254740991
                            }
                          ]
                        },
                        "parentSignedWaiverId": {
                          "nullable": true,
                          "anyOf": [
                            {
                              "type": "string",
                              "minLength": 1
                            },
                            {
                              "type": "integer",
                              "minimum": -9007199254740991,
                              "maximum": 9007199254740991
                            }
                          ]
                        },
                        "waiverId": {
                          "anyOf": [
                            {
                              "type": "string",
                              "minLength": 1
                            },
                            {
                              "type": "integer",
                              "minimum": -9007199254740991,
                              "maximum": 9007199254740991
                            }
                          ]
                        },
                        "firstName": {
                          "type": "string",
                          "minLength": 1
                        },
                        "lastName": {
                          "type": "string",
                          "minLength": 1
                        },
                        "guestId": {
                          "nullable": true,
                          "anyOf": [
                            {
                              "type": "string",
                              "minLength": 1
                            },
                            {
                              "type": "integer",
                              "minimum": -9007199254740991,
                              "maximum": 9007199254740991
                            }
                          ]
                        },
                        "dateOfBirth": {
                          "type": "string",
                          "minLength": 1
                        },
                        "email": {
                          "type": "string"
                        },
                        "contactNumber": {
                          "type": "string",
                          "minLength": 1
                        },
                        "isForMinor": {
                          "type": "boolean"
                        },
                        "expiryDate": {
                          "type": "string",
                          "minLength": 1
                        },
                        "isValid": {
                          "nullable": true,
                          "type": "boolean"
                        },
                        "createdDate": {
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": [
                        "signedWaiverId",
                        "waiverId"
                      ],
                      "additionalProperties": {}
                    }
                  }
                },
                "required": [
                  "id",
                  "sendDate",
                  "type",
                  "eventType",
                  "eventDate",
                  "data"
                ],
                "additionalProperties": {}
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Webhook accepted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "ok"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    }
  }
}